Split auth provider verifier implementation
